# Backend_CardDropper/src/main/resources/openapi/openapi.yaml
# ==============================
#   VERSION & META
# ==============================
openapi: 3.0.4
info:
  title: CardDropper API
  description: API de gestion des utilisateurs, cartes, pool et génération de packs.
  version: 1.0.0

# ==============================
#   SERVERS
# ==============================
servers:
  - url: http://localhost:8080
    description: Environnement local

# ==============================
#   TAGS
# ==============================
tags:
  - name: me
    description: Actions utilisateur authentifié (cartes, packs, échanges)
  - name: user
    description: Gestion des utilisateurs (CRUD admin + lookup)
  - name: card
    description: Gestion des cartes (CRUD + recherche)
  - name: admin
    description: Administration (templates, drop rates, gestion utilisateurs)

x-tagGroups:
  - name: Me
    tags: [ me ]
  - name: Users
    tags: [ user ]
  - name: Cards
    tags: [ card ]
  - name: Admin
    tags: [ admin ]

# ==============================
#   PATHS (ENDPOINTS)
# ==============================
paths:

  # ========================================
  #   ME (authenticated user actions)
  # ========================================
  /me/cards/owned:
    get:
      tags: [ me ]
      summary: Mes cartes possédées
      operationId: getMyCardsOwned
      responses:
        '200':
          description: Liste de cartes possédées.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardDto'
        '500':
          description: Erreur interne.

  /me/cards/created:
    get:
      tags: [ me ]
      summary: Mes cartes créées
      operationId: getMyCardsCreated
      responses:
        '200':
          description: Liste de cartes créées.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardDto'
        '500':
          description: Erreur interne.

  /me/cards/targeting:
    get:
      tags: [ me ]
      summary: Cartes me ciblant
      operationId: getMyCardsTargeting
      responses:
        '200':
          description: Liste de cartes ciblant l'utilisateur.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardDto'
        '500':
          description: Erreur interne.

  /me/cards:
    post:
      tags: [ me ]
      summary: Créer une carte (je suis le créateur)
      description: |
        Crée une nouvelle carte dans le pool global (userId = null).
        Le créateur est l'utilisateur authentifié (createdBy = moi).
        targetUserId peut être null ou égal à moi (sinon 400).
      operationId: createMyCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardDto'
      responses:
        '201':
          description: Carte créée et ajoutée au pool.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Données invalides.
        '500':
          description: Erreur interne.

  /me/cards/trade:
    post:
      tags: [ me ]
      summary: Échanger deux cartes entre moi et un autre utilisateur
      operationId: tradeMyCards
      parameters:
        - name: myCardId
          in: query
          required: true
          schema: { type: integer, format: int64 }
          example: 10
        - name: targetUserId
          in: query
          required: true
          schema: { type: integer, format: int64 }
          example: 2
        - name: theirCardId
          in: query
          required: true
          schema: { type: integer, format: int64 }
          example: 99
      responses:
        '204':
          description: Échange effectué.
        '400':
          description: Données invalides (carte ne appartenant pas à l'utilisateur, etc.).
        '404':
          description: User ou Card introuvable.
        '500':
          description: Erreur interne.

  /me/packs/open:
    post:
      tags: [ me ]
      summary: Ouvrir un pack (assigner des cartes du pool)
      description: |
        Assigne à l'utilisateur les cartes du pool (cartes dont user=null).
        Si une carte est déjà possédée -> 409.
      operationId: openMyPack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int64
            example: [ 10, 11, 12 ]
      responses:
        '200':
          description: Cartes attribuées à l'utilisateur.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardDto'
        '404':
          description: Carte introuvable.
        '409':
          description: Conflit (carte déjà possédée, donc pas dans le pool).
        '500':
          description: Erreur interne.

  /me/cards/merge:
    post:
      tags: [ me ]
      summary: Fusionner 3 cartes pour obtenir une carte de rareté supérieure
      description: |
        Règles :
          - COMMON x3 -> RARE
          - RARE x3 -> EPIC
          - EPIC x3 -> LEGENDARY
          - LEGENDARY : impossible
        Les cartes consommées retournent au pool (user=null).
        La carte résultat est tirée dans le pool à rareté supérieure.
      operationId: mergeMyCards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int64
            example: [ 10, 11, 12 ]
      responses:
        '200':
          description: Carte résultante de la fusion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Règles de fusion non respectées.
        '404':
          description: Carte introuvable.
        '409':
          description: Conflit (pool vide à rareté supérieure, ou autre conflit métier).
        '500':
          description: Erreur interne.

  /me/packs/generate:
    post:
      tags: [ me ]
      summary: Générer un pack à partir d'un template
      description: |
        Génère un pack selon le template (slots) puis assigne les cartes à l'utilisateur authentifié.
        Met à jour les dropRates des cartes non-uniques tirées.
      operationId: generateMyPack
      parameters:
        - name: templateId
          in: query
          required: true
          schema: { type: integer, format: int64 }
          example: 5
      responses:
        '200':
          description: Pack généré (cartes attribuées).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardDto'
        '404':
          description: Template introuvable.
        '409':
          description: Conflit (pool insuffisant, etc.).
        '500':
          description: Erreur interne.

  # ========================================
  #   USERS (admin CRUD + lookup)
  # ========================================
  /users:
    get:
      tags: [ user ]
      summary: Liste des utilisateurs
      description: Retourne tous les utilisateurs.
      operationId: getUsers
      responses:
        '200':
          description: Liste d'utilisateurs retournée.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '500':
          description: Erreur interne.

    post:
      tags: [ user ]
      summary: Créer un utilisateur
      description: Crée un nouvel utilisateur (username et email uniques).
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
      responses:
        '201':
          description: Utilisateur créé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Données invalides (username/email déjà pris, etc.).
        '500':
          description: Erreur interne.

  /users/{userId}:
    put:
      tags: [ user ]
      summary: Mettre à jour un utilisateur
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
      responses:
        '200':
          description: Utilisateur mis à jour.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: Utilisateur introuvable.
        '400':
          description: Données invalides.
        '500':
          description: Erreur interne.

    delete:
      tags: [ user ]
      summary: Supprimer un utilisateur
      description: Supprime un utilisateur (détache d'abord ses relations vers Card pour éviter les FK).
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '204':
          description: Utilisateur supprimé.
        '404':
          description: Utilisateur introuvable.
        '500':
          description: Erreur interne.

  /users/by-username:
    get:
      tags: [ user ]
      summary: Trouver un utilisateur par username
      operationId: getUserByUsername
      parameters:
        - name: username
          in: query
          required: true
          schema: { type: string }
          example: "andi"
      responses:
        '200':
          description: Utilisateur trouvé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: Utilisateur introuvable.
        '500':
          description: Erreur interne.

  /users/by-email:
    get:
      tags: [ user ]
      summary: Trouver un utilisateur par email
      operationId: getUserByEmail
      parameters:
        - name: email
          in: query
          required: true
          schema: { type: string, format: email }
          example: "andi@mail.com"
      responses:
        '200':
          description: Utilisateur trouvé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: Utilisateur introuvable.
        '500':
          description: Erreur interne.

  # ========================================
  #   CARDS (CRUD + QUERIES)
  # ========================================
  /cards:
    get:
      tags: [ card ]
      summary: Liste des cartes
      operationId: getCards
      responses:
        '200':
          description: Liste de cartes retournée.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardDto'
        '500':
          description: Erreur interne.

    post:
      tags: [ card ]
      summary: Créer une carte
      description: |
        Crée une carte.
        Si targetUserId est fourni, il doit être égal à createdById (sinon 400).
      operationId: createCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardDto'
      responses:
        '201':
          description: Carte créée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Données invalides.
        '500':
          description: Erreur interne.

  /cards/{cardId}:
    get:
      tags: [ card ]
      summary: Détail d'une carte
      operationId: getCardById
      parameters:
        - $ref: '#/components/parameters/CardIdPath'
      responses:
        '200':
          description: Carte trouvée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDto'
        '404':
          description: Carte introuvable.
        '500':
          description: Erreur interne.

    put:
      tags: [ card ]
      summary: Mettre à jour une carte
      operationId: updateCard
      parameters:
        - $ref: '#/components/parameters/CardIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardDto'
      responses:
        '200':
          description: Carte mise à jour.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDto'
        '404':
          description: Carte introuvable.
        '400':
          description: Données invalides.
        '500':
          description: Erreur interne.

    delete:
      tags: [ card ]
      summary: Supprimer une carte
      operationId: deleteCard
      parameters:
        - $ref: '#/components/parameters/CardIdPath'
      responses:
        '204':
          description: Carte supprimée.
        '404':
          description: Carte introuvable.
        '500':
          description: Erreur interne.

  /cards/by-rarity:
    get:
      tags: [ card ]
      summary: Rechercher des cartes par rareté
      operationId: getCardsByRarity
      parameters:
        - name: rarity
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Rarity'
          example: COMMON
      responses:
        '200':
          description: Liste de cartes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardDto'
        '400':
          description: Paramètre invalides.
        '500':
          description: Erreur interne.

  # ========================================
  #   ADMIN — PACK TEMPLATES
  # ========================================
  /admin/pack-templates:
    get:
      tags: [ admin ]
      summary: Liste des pack templates
      operationId: getPackTemplates
      responses:
        '200':
          description: Liste de templates retournée.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PackTemplateDto'
        '500':
          description: Erreur interne.

    post:
      tags: [ admin ]
      summary: Créer un pack template
      operationId: createPackTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackTemplateDto'
      responses:
        '201':
          description: Template créé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackTemplateDto'
        '400':
          description: Données invalides.
        '500':
          description: Erreur interne.

  /admin/pack-templates/{templateId}:
    get:
      tags: [ admin ]
      summary: Détail d'un pack template
      operationId: getPackTemplateById
      parameters:
        - $ref: '#/components/parameters/TemplateIdPath'
      responses:
        '200':
          description: Template trouvé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackTemplateDto'
        '404':
          description: Template introuvable.
        '500':
          description: Erreur interne.

    put:
      tags: [ admin ]
      summary: Mettre à jour un pack template
      operationId: updatePackTemplate
      parameters:
        - $ref: '#/components/parameters/TemplateIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackTemplateDto'
      responses:
        '200':
          description: Template mis à jour.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackTemplateDto'
        '404':
          description: Template introuvable.
        '400':
          description: Données invalides.
        '500':
          description: Erreur interne.

    delete:
      tags: [ admin ]
      summary: Supprimer un pack template
      operationId: deletePackTemplate
      parameters:
        - $ref: '#/components/parameters/TemplateIdPath'
      responses:
        '204':
          description: Template supprimé.
        '404':
          description: Template introuvable.
        '500':
          description: Erreur interne.

  # ========================================
  #   ADMIN — DROP RATES
  # ========================================
  /admin/drop-rates/by-rarity:
    put:
      tags: [ admin ]
      summary: Modifier le dropRate de toutes les cartes du pool pour une rareté
      description: Met à jour le dropRate de toutes les cartes dans le pool (user=null) ayant la rareté spécifiée.
      operationId: updateDropRateByRarity
      parameters:
        - name: rarity
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Rarity'
          example: LEGENDARY
        - name: dropRate
          in: query
          required: true
          schema:
            type: number
            format: double
          example: 0.02
      responses:
        '200':
          description: Nombre de cartes mises à jour.
          content:
            application/json:
              schema:
                type: integer
                format: int32
        '400':
          description: Paramètres invalides.
        '500':
          description: Erreur interne.

  # ========================================
  #   ADMIN — USERS
  # ========================================
  /admin/users:
    get:
      tags: [ admin ]
      summary: Liste de tous les utilisateurs (vue admin)
      description: Retourne tous les utilisateurs avec leurs cartes possédées, créées et ciblées.
      operationId: getAdminUsers
      responses:
        '200':
          description: Liste d'utilisateurs retournée.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '500':
          description: Erreur interne.

  /admin/users/{userId}:
    get:
      tags: [ admin ]
      summary: Détail d'un utilisateur (vue admin)
      operationId: getAdminUserById
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Utilisateur trouvé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: Utilisateur introuvable.
        '500':
          description: Erreur interne.

# ==============================
#   COMPONENTS
# ==============================
components:

  # ------------------------------
  # Parameters
  # ------------------------------
  parameters:
    UserIdPath:
      name: userId
      in: path
      required: true
      description: Identifiant de l'utilisateur.
      schema:
        type: integer
        format: int64
      example: 1

    CardIdPath:
      name: cardId
      in: path
      required: true
      description: Identifiant de la carte.
      schema:
        type: integer
        format: int64
      example: 10

    TemplateIdPath:
      name: templateId
      in: path
      required: true
      description: Identifiant du pack template.
      schema:
        type: integer
        format: int64
      example: 5

  # ------------------------------
  # Schemas
  # ------------------------------
  schemas:

    # --- Enums ---
    Rarity:
      type: string
      description: Rareté d'une carte.
      enum: [ COMMON, RARE, EPIC, LEGENDARY ]
      example: COMMON

    # --- DTOs ---
    CardDto:
      type: object
      description: Représentation d'une carte.
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: "Dragon"
        imageUrl:
          type: string
          example: "https://img.example.com/cards/dragon.png"
        rarity:
          $ref: '#/components/schemas/Rarity'
        description:
          type: string
          example: "Carte légendaire très rare."
        dropRate:
          type: number
          format: double
          example: 0.12
        uniqueCard:
          type: boolean
          example: false
        userId:
          type: integer
          format: int64
          nullable: true
          description: Propriétaire actuel (null = dans le pool).
          example: 1
        createdById:
          type: integer
          format: int64
          nullable: true
          description: Créateur de la carte.
          example: 1
        targetUserId:
          type: integer
          format: int64
          nullable: true
          description: Utilisateur ciblé (optionnel).
          example: 1
      required: [ name, rarity, dropRate, uniqueCard ]
      example:
        id: 10
        name: "Dragon"
        imageUrl: "https://img.example.com/cards/dragon.png"
        rarity: "EPIC"
        description: "Une carte puissante."
        dropRate: 0.08
        uniqueCard: false
        userId: null
        createdById: 1
        targetUserId: null

    UserDto:
      type: object
      description: Représentation d'un utilisateur.
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: "andi"
        email:
          type: string
          format: email
          example: "andi@mail.com"
        passwordHash:
          type: string
          description: Hash du mot de passe (si exposé dans l'API).
          example: "$2a$10$something..."
        cardsOwned:
          type: array
          items:
            $ref: '#/components/schemas/CardDto'
        cardsCreated:
          type: array
          items:
            $ref: '#/components/schemas/CardDto'
        cardsTargeting:
          type: array
          items:
            $ref: '#/components/schemas/CardDto'
      required: [ username, email ]
      example:
        id: 1
        username: "andi"
        email: "andi@mail.com"
        passwordHash: "$2a$10$something..."
        cardsOwned: [ ]
        cardsCreated: [ ]
        cardsTargeting: [ ]

    PackTemplateDto:
      type: object
      description: Template de pack définissant les slots et leurs probabilités.
      properties:
        id:
          type: integer
          format: int64
          example: 5
        name:
          type: string
          example: "Standard Pack"
        slots:
          type: array
          items:
            $ref: '#/components/schemas/PackSlotDto'
      required: [ name ]
      example:
        id: 5
        name: "Standard Pack"
        slots:
          - fixedRarity: "RARE"
          - rarityWeights:
              COMMON: 70.0
              RARE: 25.0
              EPIC: 4.0
              LEGENDARY: 1.0

    PackSlotDto:
      type: object
      description: Slot d'un pack template (rareté fixe ou pondérée).
      properties:
        id:
          type: integer
          format: int64
          example: 1
        fixedRarity:
          $ref: '#/components/schemas/Rarity'
        rarityWeights:
          type: object
          additionalProperties:
            type: number
            format: double
          description: "Poids par rareté (ex: { COMMON: 70.0, RARE: 25.0 })"
          example:
            COMMON: 70.0
            RARE: 25.0
            EPIC: 4.0
            LEGENDARY: 1.0

  # ------------------------------
  # Security
  # ------------------------------
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []